Split AsciidocSupport class into the specific handler classes
